1.添加远端共享及云端共享模块控制变量，不用是否登陆来判断是否重新加载。 2.删除局域网共享右键中无用的操作。 3.添加是否删除本地共享文件窗口提示。

+++ vlc-2.2.0/modules/gui/qt4/components/playlist/standardpanel.cpp
@@ -72,6 +72,7 @@
 #include <QUrl>
 #include <QFont>
 #include <QMessageBox>
+#include <QDebug>
 
 #include <assert.h>
 #include <dirent.h>
@@ -318,11 +319,6 @@ bool StandardPLPanel::popup( const QPoint &point )
 		ADD_MENU_ENTRY( QIcon( ":/buttons/playlist/playlist_remove" ), qtr("删除云共享文件"), VLCModelSubInterface::ACTION_DELCLOUD );
     }
 
-    if (p_selector->getCurrentItemCategory() == LANSHARE )
-    {
-        menu.addAction( qtr( "ADD LAN SHARE" ), this, SLOT( increaseZoom() ) );
-        menu.addAction( qtr( "DELETE LAN SHARE" ), this, SLOT( increaseZoom() ) );
-    }
     if (p_selector->getCurrentItemCategory() == REMOTESHARE )
     {
         //menu.addAction( qtr( "ADD REMOTE SHARE" ), this, SLOT( increaseZoom() ) );
@@ -487,31 +483,31 @@ void StandardPLPanel::popupAction( QAction *action )
 
 			/*add by lili*/
 		case VLCModelSubInterface::ACTION_DELLOCAL:
+			if (index.data().toString().toStdString().length())
 			{
-			QString file = " rm ";
-			file.append(sharePath);
-			file.append("/");
-			file.append( qtu(index.data().toString()));
-			printf("qtu index.data:%s\n", qtu(index.data().toString()));
-			printf( "del file:%s\n", file.toStdString().c_str());
-			printf("file no qtu:%s\n", file.toStdString().c_str() );
-			system( file.toStdString().c_str() );
-
-			
-#if 0		
-			char file[1024] ;
-			memset( file, 0, sizeof(file));
-			sprintf( file, "%s/", qtu(QString(sharePath)) );
-			//strcat( file, index.data().toString().toStdString().c_str() );
-			strcat( file, qtu(index.data().toString()) );
-			printf("del file %s\n", qtu(QString(file)));
-#endif
-			//QFile::remove( file.toStdString().c_str()  );
-			
-			model->action( action, list );
+				QMessageBox msgBox( QMessageBox::Information,
+					qtr( "删除文件" ),
+					qtr( "您确定要删除该本地共享文件吗？" ),
+					QMessageBox::Yes | QMessageBox::No,
+					NULL );
+				int ret =  msgBox.exec();
+				if( ret == QMessageBox::Yes )
+				{
+					QString localfile = " rm ";
+					localfile.append(sharePath);
+					localfile.append("/");
+					localfile.append( qtu(index.data().toString()));
+					printf("qtu index.data:%s\n", qtu(index.data().toString()));
+					printf( "del file:%s\n", localfile.toStdString().c_str());
+					system( localfile.toStdString().c_str() );
+					model->action( action, list );
+				}
+				else if( ret == QMessageBox::No )
+				{
+					qDebug() << "do not remove file";
+				}
 			}
 			break;
-
 		case VLCModelSubInterface::ACTION_ADDLOCAL:
         //pl_item = playlist_ChildSearchName( THEPL->p_root, qtu( item->data(0, LONGNAME_ROLE ).toString() ) );
 
@@ -903,7 +899,7 @@ void StandardPLPanel::browseInto( const QModelIndex &index )
 /* add by lili */
 void StandardPLPanel::createCloudItems( const QModelIndex &index )
 {
-	printf( "--------%s:%d------------\n", __func__, __LINE__ );
+	printf( "------------%s:%d------------\n", __func__, __LINE__ );
 	UserOption *user = UserOption::getInstance( p_intf );
 	if( user == NULL )
 	{
@@ -922,6 +918,8 @@ void StandardPLPanel::createCloudItems( const QModelIndex &index )
 		msgBox.exec();
 		return ;
 	}
+
+	user->setCloudSharedStart(true);
 
 	playlist_item_t *play_item = playlist_ItemGetById( THEPL, model->itemId( index, PLAYLIST_ID ) );
 	if( play_item == NULL || play_item->i_children > 0 )
@@ -990,6 +988,8 @@ void StandardPLPanel::createRemoteShareItems( const QModelIndex &index )
 		msgBox.exec();
 		return ;
 	}
+
+	user->setRemoteSharedStart(true);
 
 	mid = user->getLUid();
